BUILDABLE Implemented WiFiManager
Removed the original code to manage the wifi connection and implemented
the standard WiFiManager contrib module to take care of that setup.
Rearranged the project to more closely match what I would consider a
standard project (moved files around).
