--- OpenGarage/APWizard.h
@@ -1,56 +1,0 @@
-/* OpenGarage Firmware
- *
- * ESPConnect header file
- * Mar 2016 @ OpenGarage.io
- *
- * This file is part of the OpenGarage library
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see
- * <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _AP_WIZARD_H
-#define _AP_WIZARD_H
-
-#define HTML_OK                0x00
-#define HTML_SUCCESS           0x01
-#define HTML_UNAUTHORIZED      0x02
-#define HTML_MISMATCH          0x03
-#define HTML_DATA_MISSING      0x10
-#define HTML_DATA_OUTOFBOUND   0x11
-#define HTML_DATA_FORMATERROR  0x12
-#define HTML_PAGE_NOT_FOUND    0x20
-#define HTML_FILE_NOT_FOUND    0x21
-#define HTML_NOT_PERMITTED     0x30
-#define HTML_UPLOAD_FAILED     0x40
-#define HTML_REDIRECT_HOME     0xFF
-
-#include <ESP8266WiFi.h>
-#include <ESP8266WebServer.h>
-#include <ESP8266mDNS.h>
-#include <ArduinoJson.h>
-#include <WiFiUdp.h>
-#include <time.h>
-#include "defines.h"
-
-String scan_network();
-void start_network_ap(const char *ssid, const char *pass);
-void start_network_sta(const char *ssid, const char *pass);
-void start_network_sta_with_ap(const char *ssid, const char *pass);
-
-uint curr_utc_time();
-
-extern const char html_ap_home[] PROGMEM;
-extern const char html_portal[] PROGMEM;
-#endif
