+++ OpenGarage/Assets/compile_assets.sh
@@ -1,0 +1,2 @@
+rm compiled.h
+./compile_assets portal.html html_portal
