+++ OpenGarage/OpenGarage-Arduino/OpenGarage-Arduino.ino
@@ -28,7 +28,7 @@
 #include <WiFiUdp.h>
 #include <time.h>
 #include <FS.h>
-#include "OpenGarage.h"
+#include "../main.cpp";
 
 void do_setup();
 void do_loop();
