+++ OpenGarage/defines.h
@@ -20,6 +20,15 @@
  * <http://www.gnu.org/licenses/>.
  */
 
+#include <ESP8266WiFi.h>
+#include <ESP8266WebServer.h>
+#include <ESP8266mDNS.h>
+#include <ArduinoJson.h>
+#include <WiFiManager.h>
+#include <WiFiUdp.h>
+#include <time.h>
+#include <FS.h>
+
 #ifndef _DEFINES_H
 #define _DEFINES_H
 
@@ -67,6 +76,19 @@
 #define DOOR_STATUS_JUST_OPENED   2
 #define DOOR_STATUS_JUST_CLOSED   3
 #define DOOR_STATUS_MIXED         4
+
+#define HTML_OK                0x00
+#define HTML_SUCCESS           0x01
+#define HTML_UNAUTHORIZED      0x02
+#define HTML_MISMATCH          0x03
+#define HTML_DATA_MISSING      0x10
+#define HTML_DATA_OUTOFBOUND   0x11
+#define HTML_DATA_FORMATERROR  0x12
+#define HTML_PAGE_NOT_FOUND    0x20
+#define HTML_FILE_NOT_FOUND    0x21
+#define HTML_NOT_PERMITTED     0x30
+#define HTML_UPLOAD_FAILED     0x40
+#define HTML_REDIRECT_HOME     0xFF
 
 typedef enum {
 	OPTION_FIRMWARE_VERSION = 0, // firmware version
