+++ OpenGarage/libraries/ConfigFile.cpp
