+++ OpenGarage/libraries/Logging.cpp
