+++ OpenGarage/libraries/Logging.h
