--- OpenGarage/www_assets/ap_home.html
@@ -1,71 +1,0 @@
-<head>
-<meta name='viewport' content='width=device-width,initial-scale=1.0,minimum-scale=1.0,user-scalable=no'>
-<title>OpenGarage</title>
-<style>body{font-family:'helvetica';}</style>
-</head>
-<body>
-	<caption><b>OpenGarage WiFi Config</caption><br><br>
-<table cellspacing=4 id='rd'>
-<tr><td>(Scanning...)</td></tr>
-</table>
-<table cellspacing=16>
-<tr><td><input type='text' name='ssid' id='ssid'></td><td>(SSID)</td></tr>
-<tr><td><input type='password' name='pass' id='pass'></td><td>(Password)</td></tr>
-<tr><td colspan=2><p id='msg'></p></td></tr>
-<tr><td><button type='button' id='butt' onclick='submit_form();' style='height:36px;width:180px'>Submit</button></td><td></td></tr>
-</table>
-<script>
-	function id(s) {return document.getElementById(s);}
-	function sel(i) {id('ssid').value=id('rd'+i).value;}
-	var tci;
-	function tryConnect() {
-		var xhr=new XMLHttpRequest();
-		xhr.onreadystatechange=function() {
-			if(xhr.readyState==4 && xhr.status==200) {
-				var jd=JSON.parse(xhr.responseText);
-				if(jd.ip==0) return;
-				var ip=''+(jd.ip%256)+'.'+((jd.ip/256>>0)%256)+'.'+(((jd.ip/256>>0)/256>>0)%256)+'.'+(((jd.ip/256>>0)/256>>0)/256>>0);
-				id('msg').innerHTML='<b><font color=green>Connected! Device IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'
-				id('butt').innerHTML='Go to '+ip;id('butt').disabled=false;
-				id('butt').onclick=function rd(){window.open('http://'+ip);}
-				clearInterval(tci);
-			}
-		}    
-		xhr.open('POST', 'jt', true); xhr.send();    
-	}
-	function submit_form() {
-		id('msg').innerHTML='';
-		var xhr=new XMLHttpRequest();
-		xhr.onreadystatechange=function() {
-			if(xhr.readyState==4 && xhr.status==200) {
-				var jd=JSON.parse(xhr.responseText);
-				if(jd.result==1) { tci=setInterval(tryConnect, 10000); return; }
-				id('msg').innerHTML='<b><font color=red>Error code: '+jd.result+', item: '+jd.item+'</font></b>';
-				id('butt').innerHTML='Submit';
-				id('butt').disabled=false;id('ssid').disabled=false;id('pass').disabled=false;
-			}
-		}
-		var comm='cc?ssid='+encodeURIComponent(id('ssid').value)+'&pass='+encodeURIComponent(id('pass').value);
-		xhr.open('POST', comm, true); xhr.send();
-		id('butt').innerHTML='Connecting...';
-		id('butt').disabled=true;id('ssid').disabled=true;id('pass').disabled=true;
-	}
-		
-	function loadSSIDs() {
-		var xhr=new XMLHttpRequest();
-		xhr.onreadystatechange=function() {
-			if(xhr.readyState==4 && xhr.status==200) {
-				var response=JSON.parse(xhr.responseText), i;
-				id('rd').deleteRow(0);
-				for(i=0;i<response.networks.length;i++) {
-					var cell=id('rd').insertRow(-1).insertCell(0);
-					cell.innerHTML ="<tr><td><input type='radio' name='ssids' id='rd" + i + "' value='" + response.networks[i].ssid + "' onclick=sel(" + i + ")>" + response.networks[i].ssid + "(" + response.networks[i].rssi + ")" + "</td></tr>";
-
-				}
-			}
-		};
-		xhr.open('GET','/json/networks',true); xhr.send();
-	}
-	setTimeout(loadSSIDs, 1000);
-</script>
-</body>
