--- OpenGarage/www_assets/publish_assets.sh
@@ -1,3 +1,0 @@
-rm ../www_assets.h
-./publish_assets ap_home.html html_ap_home
-./publish_assets portal.html html_portal
